Prepare web worker for AWS (#105)
* Prepare web worker for AWS

This PR adds a `/health` endpoint that does the same thing as
`/health/ping` to make ALB health checks work. It also adds a toggle for
disabling HTTPS which is not supported between ALBs and EC2 instances.

* Update circle config to work with ECS specific branches

To keep the branch history clean, we're going to use dedicated ECS
branches (prefixed with `ecs_`) for new infra related commits. This PR
updates the circle config to take this into account.